requirements.txt
django
django-rest-framework
drf-spectacular
djangorestframework-simplejwt
Pillow
djangorestframework
